Tidy up the new Windows MIDI code a bit

+++ src/m_misc.h
@@ -74,7 +74,7 @@ const char *M_StrCaseStr(const char *haystack, const char *needle);
 void M_StringCopy(char *dest, const char *src, const size_t dest_size);
 char *M_StringReplace(char *haystack, char *needle, const char *replacement);
 void M_StringReplaceAll(char *haystack, char *needle, const char *replacement);
-dboolean M_WriteFile(const char *name, const void *source, int length);
+dboolean M_WriteFile(const char *name, const void *source, size_t length);
 char *M_TempFile(char *s);
 char *M_StringJoin(char *s, ...);
 dboolean M_StringStartsWith(const char *s, const char *prefix);
